requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
discord
python-dotenv
detoxify
tensorflow-cpu==1.15
tensorflow-hub==0.7.0
pillow
